hashtag3/scripts/requirements.txt
ImageFilter == 0.1.0
Pillow == 8.3.2
numpy == 1.21.2
scipy == 1.7.1
backports.entry-points-selectable == 1.1.0
distlib == 0.3.3
filelock == 3.3.0
platformdirs == 2.4.0